# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Stayinn API",
    "version": "1.0.0",
    "description": "API Specification for Stayinn — Hotel Booking Marketplace"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local Development Server"
    },
    {
      "url": "https://stayinn.ng/api",
      "description": "Production Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/hotels": {
      "get": {
        "summary": "List approved hotels",
        "tags": ["Hotels"],
        "security": [],
        "responses": {
          "200": { "description": "List of hotels" }
        }
      },
      "post": {
        "summary": "Create a hotel",
        "tags": ["Hotels"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "address": { "type": "string" },
                  "latitude": { "type": "number" },
                  "longitude": { "type": "number" },
                  "coverImage": { "type": "string" },
                  "amenities": {
                    "type": "array",
                    "items": { "type": "string", "enum": ["WIFI", "WATER", "BACKUP_POWER", "PARKING", "POOL", "GYM", "RESTAURANT", "AIR_CONDITIONING"] }
                  }
                },
                "required": ["name", "address", "latitude", "longitude", "coverImage"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Hotel created (pending approval)" }
        }
      }
    },
    "/hotels/{id}": {
      "get": {
        "summary": "Get hotel details",
        "tags": ["Hotels"],
        "security": [],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Hotel details" }
        }
      },
      "put": {
        "summary": "Update hotel details",
        "tags": ["Hotels"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Hotel updated successfully" }
        }
      }
    },
    "/hotels/{id}/rooms": {
      "get": {
        "summary": "List active room types for a hotel",
        "tags": ["Rooms"],
        "security": [],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "List of room types" }
        }
      },
      "post": {
        "summary": "Create a room type for a hotel",
        "tags": ["Rooms"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "pricePerNight": { "type": "number" },
                  "capacity": { "type": "integer" },
                  "quantity": { "type": "integer" },
                  "images": { "type": "array", "items": { "type": "string" } }
                },
                "required": ["name", "pricePerNight", "capacity", "quantity", "images"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Room type created successfully" }
        }
      }
    },
    "/hotels/{id}/rooms/{roomId}": {
      "get": {
        "summary": "Get room details",
        "tags": ["Rooms"],
        "security": [],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "roomId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Room type details" }
        }
      },
      "put": {
        "summary": "Update room details",
        "tags": ["Rooms"],
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "roomId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Room updated successfully" }
        }
      }
    },
...
    "/payments/initiate": {
      "post": {
        "summary": "Initiate Paystack Payment for a booking",
        "tags": ["Payments"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookingId": { "type": "string" }
                },
                "required": ["bookingId"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Payment authorization URL" }
        }
      }
    },
...
  }
}